refactor placement + compile_shaders and big performance improvements
- caching parameters
- caching source sha256 hashes.
- caching glsl compilations (not 100% done yet)

+++ database/src/denox/db/Db.cpp
@@ -204,7 +204,7 @@ denox::Db::query_shader_binary(const SHA256 &srcHash) const {
   if (it == m_index->binary_index.end()) {
     return std::nullopt;
   } else {
-    uint32_t binaryId = it->second;
+    uint32_t binaryId = static_cast<uint32_t>(it->second);
     return m_db->binaries[binaryId].spvBinary;
   }
 }
